Refactor timing and memory management in DIMSE operations
Introduce an elapsed timer for measuring the duration of DIMSE operations, ensuring accurate reporting even when the host clock does not advance. Update the PDataReader to manage fragmented data more efficiently, reducing memory overhead by retaining references to received fragments instead of copying data. Enhance the CEcho verification process to utilize the new timer, providing clearer duration metrics for both successful and failed operations. Comprehensive tests are added to validate the new timing functionality and memory management improvements.

+++ net/dimse/encode.go
@@ -25,7 +25,7 @@ func EncodeCommandSet(elements []core.Element) ([]byte, error) {
 	if err := object.WriteDataSet(&command, obj, transfer.ImplicitVRLittleEndian); err != nil {
 		return nil, fmt.Errorf("dicom dimse: encode command set: %w", err)
 	}
-	if command.Len() > math.MaxUint32 {
+	if uint64(command.Len()) > math.MaxUint32 {
 		return nil, fmt.Errorf("dicom dimse: command group length %d exceeds uint32", command.Len())
 	}
